[dhl88] new logs preping for long scan.  analyzing the daqspectrum image
Preping for the next long scan

Wrote script daqspectrum_image to help analyze the daqspectrum image
taken overnight.

An easier method would be to rewrite the longscan_daqspectrum helper
script to store the location being scanned at and daqspectrum for
each position.  Have to see how large that would be...